+++ platform/migrations/0008_slug_held_until.sql
@@ -1,0 +1,1 @@
+ALTER TABLE site_billing ADD COLUMN slug_held_until INTEGER;